Add maintenance process kb
A customer was asking about the vulnerability and patching concept at
APPUiO Public. We do not have written down our maintenance process yet.
The couple of sentences which were written for the customer are now
documented here and could be extended. Originally it was in our wiki in
VINT space with name "Maintenance Process".