ISMS-631/VT-1072 Add Security Process (#38)
For our ISAE 3402 report control 5.6.1 we want to document our
vulnerability process. This is highly related to our security process
ticket VT-1072 which we wanted to tackle down long time ago.

This commit adds the process as it is currently handled in our company.

Improvements of the process can now be documented and discussed here and
ensured with trainings of our employees later.